Rename file with test generators
